requirements.txt
sentence-transformers==3.0.1
lancedb==0.13.0
langchain-core==0.3.63
langchain-google-genai==2.0.5
streamlit==1.38.0
python-dotenv==1.0.1
pandas==2.2.2
pydantic==2.8.2
torch==2.4.0
transformers==4.44.0
numpy==1.26.4
pypdf==6.0.0
python-docx==1.1.2
google-generativeai==0.8.3
